# app/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Library Management API
  version: 1.0.0
paths:
  /books:
    get:
      summary: List all books
      responses:
        '200':
          description: A list of all books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, author, published_year, isbn]
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                isbn:
                  type: string
                genre:
                  type: string
      responses:
        '201':
          description: Book created successfully
  /books/search:
    get:
      summary: Search for books by filters
      parameters:
        - in: query
          name: author
          schema:
            type: string
        - in: query
          name: published_year
          schema:
            type: integer
        - in: query
          name: genre
          schema:
            type: string
      responses:
        '200':
          description: Filtered list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /books/{isbn}:
    delete:
      summary: Delete a book by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book deleted successfully
        '404':
          description: Book not found
    put:
      summary: Update a book by ISBN
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                genre:
                  type: string
      responses:
        '200':
          description: Book updated successfully
        '404':
          description: Book not found